Update fit() docstring
Shortened overall introduction.

Expaned explanation for use_cut_trials argument, and added warning
that it might lead to worse results if long-term dependencies
exist in the data. Removed the detailed description of
use_cut_trials in Notes, as the user of fit() has no way to change
the argument that is provided to _cut_trials, making any notes
of this argument irrelevant to the user of fit().

Thightened the explanation for how the EM algorithm is implemented
by this function.